added event on  command and fixed some tests
